Fix problem that .bashrc does not load if mac

+++ init/setup/link.sh
@@ -4,5 +4,9 @@ mkdir -p ~/.config/fish
 
 # link dotfiles
 ln -sf ~/dotfiles/.tmux.conf ~/.tmux.conf
-ln -sf ~/dotfiles/.bashrc ~/.bashrc
 ln -sf ~/dotfiles/fish/config.fish  ~/.config/fish/config.fish
+if [ "$(uname)" == 'Darwin' ]; then
+  ln -sf ~/dotfiles/.bashrc ~/.bash_profile
+elif [ -e /etc/debian_version ]; then
+  ln -sf ~/dotfiles/.bashrc ~/.bashrc
+fi
